\name{print.StabilityCollection}
\alias{print.StabilityCollection}
\title{Print a brief summary of the stability of a clustering collection.}
\usage{
  \method{print}{StabilityCollection} (clusterings)
}
\description{
  Print a brief summary of the stability of a clustering
  collection.
}
\seealso{
  \code{\link{easystab}}
}

