\name{print.StabilityReport}
\alias{print.StabilityReport}
\title{Print a brief summary of the stability of an undividual clustering under perturbation.}
\usage{
  \method{print}{StabilityReport} (clustering)
}
\description{
  Print a brief summary of the stability of an undividual
  clustering under perturbation.
}

