\name{summary.StabilityCollection}
\alias{summary.StabilityCollection}
\title{Print a detaild summary of the stability of a clustering collection.}
\usage{
  \method{summary}{StabilityCollection} (clusterings)
}
\description{
  Print a detaild summary of the stability of a clustering
  collection.
}

