\name{summary.StabilityReport}
\alias{summary.StabilityReport}
\title{Print a summary of the stability of an undividual clustering under perturbation.}
\usage{
  \method{summary}{StabilityReport} (clustering)
}
\description{
  Print a summary of the stability of an undividual
  clustering under perturbation.  Summary includes
  individual cluster stabilites.
}

