\name{plot.StabilitySequence}
\alias{plot.StabilitySequence}
\title{Plot the stability scores produced by perturbationStability as a sequence of
box plots.}
\usage{
  plot.StabilitySequence(clusterings)
}
\arguments{
  \item{clusterings}{The output of
  \code{perturbationStablity} -- a list of clusters with
  perturbation stability analyses.}
}
\description{
  Summary display of the output of the
  perturbationStability function. Plot the stability scores
  produced by perturbationStability as a sequence of box
  plots.
}

