\name{plotLabeledStabilityMap}
\alias{plotLabeledStabilityMap}
\title{plot stability map with labels based on clustering stability analysis results}
\description{perturbationStability function post processing. plot
  stability map with labels based on clustering stability analysis results}
\usage{
plotLabeledStabilityMap(clustering)
}
\arguments{
  \item{clustering}{A list object of a clustering with stability analysis results attached}
}
\seealso{
  \code{\link{easystab}}, \code{\link{perturbationStability}},
  \code{\link{plotStabilityMap}}
}

