\name{plotStabilityImage}

\alias{plotStabilityImage}

\title{plot pointwise stability image}

\description{plot pointwise stability image}

\usage{
plotStabilityImage(centroids, theta, image_nx, image_ny, image_x_lower = 0, image_y_upper = 0, image_y_lower = 0, image_y_upper =0)
}

\arguments{
  \item{centroids}{matrix of centroid points}
  
  \item{theta}{theta}
  
  \item{image_nx}{image pixel number in x direction}
  
  \item{image_ny}{image pixel number in y direction}
  
  \item{image_x_lower}{minimum x value}
  
  \item{image_x_upper}{maximum x value}
  
  \item{image_y_lower}{minimum y value}
  
  \item{image_y_upper}{maximum y value}
}