\name{plotStabilityMap}
\alias{plotStabilityMap}
\title{plot stability map based on clustering stability analysis results}
\description{perturbationStability function post processing. plot stability map based on clustering stability analysis results}
\usage{
plotStabilityMap(clustering)
}
\arguments{
  \item{clustering}{A list object of a clustering with stability analysis results attached}
}

