\name{plotStabilitySequence}
\alias{plotStabilitySequence}
\title{plot stability sequence based on clustering stability analysis results}
\description{perturbationStability function post processing. plot stability sequence based on clustering stability analysis results}
\usage{
plotStablilitySequence(clusterings)
}
\arguments{
  \item{clusterings}{A list of clusters with perturbation stability
    analysis results. It is the result from perturbationStablity function}
}

