\name{print.StabilityCollection}
\alias{print.StabilityCollection}
\title{Print a brief summary of the stability of a clustering collection.}
\usage{
  \method{print}{StabilityCollection} (x, ...)
}
\arguments{
  \item{x}{The output of \code{perturbationStablity} -- a
  list of clusters with perturbation stability analyses.}

  \item{...}{optional arguments passed to internal
  functions}
}
\description{
  Print a brief summary of the stability of a clustering
  collection.
}
\seealso{
  \code{\link{easystab}}
}

