\name{print.StabilityReport}
\alias{print.StabilityReport}
\title{Print a brief summary of the stability of an undividual clustering under perturbation.}
\usage{
  \method{print}{StabilityReport} (x, ...)
}
\arguments{
  \item{x}{A StabilityReport object, as given by an output
  of perturbationStability.}

  \item{...}{optional arguments passed to internal
  functions}
}
\description{
  Print a brief summary of the stability of an undividual
  clustering under perturbation.
}
\seealso{
  \code{\link{easystab}}
}

