\name{summary.StabilityCollection}
\alias{summary.StabilityCollection}
\title{Print a detaild summary of the stability of a clustering collection.}
\usage{
  \method{summary}{StabilityCollection} (object, ...)
}
\arguments{
  \item{object}{The output of \code{perturbationStablity}
  -- a list of clusters with perturbation stability
  analyses.}

  \item{...}{optional argements passed to internal
  functions}
}
\description{
  Print a detaild summary of the stability of a clustering
  collection.
}
\seealso{
  \code{\link{easystab}}
}

