\name{summary.StabilityReport}
\alias{summary.StabilityReport}
\title{Print a summary of the stability of an undividual clustering under perturbation.}
\usage{
  \method{summary}{StabilityReport} (object, ...)
}
\arguments{
  \item{object}{A StabilityReport object, as given by an
  output of perturbationStability.}

  \item{...}{optional arguments passed to internal
  functions}
}
\description{
  Print a summary of the stability of an undividual
  clustering under perturbation.  Summary includes
  individual cluster stabilites.
}
\seealso{
  \code{\link{easystab}}
}

